pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[project]
name = "phototags"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "Pillow>=10.0",
    "PyYAML>=6.0",
    "rawpy>=0.19",
    "numpy>=1.20",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
phototags = "phototags.main:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
